cleanup(8/n): rename strawberry_schema.py -> schema.py
With the legacy graphene root schema.py deleted, the canonical name is free. Rename the
Strawberry module to solvis_graphql_api/schema.py and repoint importers (app.py, the SDL
parity gate, _strawberry_client shim, the snapshot + corpus-replay tests). Module path only —
local 'as strawberry_schema' aliases kept. Refreshed the module docstring to point at the
graphene-free compute homes (cached, ruptures, color_scale.compute, geojson_style_util).

app + handler import OK; SDL byte-identical; 83 pass, 10 skipped; ruff clean.
