+++ tests/test_corpus_replay.py
@@ -12,7 +12,7 @@ import pathlib
 import pytest
 from graphql import build_schema, parse, validate
 
-from solvis_graphql_api.strawberry_schema import schema as strawberry_schema
+from solvis_graphql_api.schema import schema as strawberry_schema
 
 CORPUS_DIR = pathlib.Path(__file__).parent / "fixtures" / "corpus"
 CORPUS = sorted(CORPUS_DIR.glob("*.graphql"))
